+++ lib/handbrake.rb
@@ -8,4 +8,10 @@ module HandBrake
   autoload :Titles,  'handbrake/titles'
   autoload :Title,   'handbrake/titles'
   autoload :Chapter, 'handbrake/titles'
+
+  ##
+  # The exception thrown when a file exists that shouldn't.
+  #
+  # @see CLI#output
+  class FileExistsError < StandardError; end;
 end